Adapt to OpenSSL 1.1.1.
The FreeBSD operating system is migrating to OpenSSL 1.1.1 and I have created this set of patches to make pam_ssh_agent_auth compile with it.

The patch comments out some parts of include files which are not actually used and reference now opaque OpenSSL internals.

I also have migrated the source files to use accessors to use the OpenSSL objects.

The patch works on FreeBSD head (will be 12.0) but the --without-openssl-header-check argument is required in configure there.